% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMDVSerPred.R
\name{feat.norm}
\alias{feat.norm}
\title{Normalization of k-mer count data}
\usage{
feat.norm(dat)
}
\arguments{
\item{dat}{Dataframe of k-mer counts of the virus isolates usually obtained from kmer techniques.}
}
\value{
Data frame containing normalized feature values of k-mers, where rows are virus isolates & columns are features.
}
\description{
Normalizes the k-mer counts data to remove bias due to varying sequence lengths.
}
\author{
Samarendra Das
}
