% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMDVSerPred.R
\name{Feature.dna}
\alias{Feature.dna}
\title{Feature generation from sequence data}
\usage{
Feature.dna(x, k)
}
\arguments{
\item{x}{Nucleotide sequence data of virus isolates in fasta format.}

\item{k}{Integral number (>0) indicating the size of mers to generate features.}
}
\value{
Data frame containing feature counts of k-mers, where rows are virus isolates & columns are features.
}
\description{
Returns k-mer counts data.
}
\author{
Samarendra Das
}
