% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMDVSerPred.R
\name{Feature.sel}
\alias{Feature.sel}
\title{Releveant k-mer feature selection.}
\usage{
Feature.sel(x, serotypes, n)
}
\arguments{
\item{x}{Nucleotide sequence data of virus isolates in fasta format.}

\item{serotypes}{Serotype information of the FMD virus isolates, whose sequences are given in x}

\item{n}{Integral number (>0) indicating the number of relevant features to be included in the model.}
}
\value{
Data frame containing feature counts of k-mers, where rows are virus isolates & columns are features.
\itemize{
  \item1 feat.sel is the list of selected features.
  \item2 x.sel is the k-mer count of the virus isolates over the selected features.
 }
}
\description{
Returns the list of relevant k-mer features required for model training.
}
\author{
Samarendra Das
}
