% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMDVSerPred.R
\name{serotypPred}
\alias{serotypPred}
\title{Prediction of unknown serotypes of isolates.}
\usage{
serotypPred(x, serotyp, QuerySeq)
}
\arguments{
\item{x}{Nucleotide sequence data of virus isolates in fasta format for model training.}

\item{serotyp}{Serotype information of the virus isolates whose sequence data is given in x.}

\item{QuerySeq}{Sequence data of the virus isolates in fasta format, whose serotype will be predicted.}
}
\value{
Data frame containing probabilities for the serotypes computed for each isolate and right-most column contains predicted serotypes.
}
\description{
Returns the serotypes of the novel FMD virus isolates supplied in query sequence data.
}
\author{
Samarendra Das
}
