% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMDVSerPred.R
\name{SerotypPredModel}
\alias{SerotypPredModel}
\title{Model training for serotype prediction and classification.}
\usage{
SerotypPredModel(x, serotyp, ind, m, k, n)
}
\arguments{
\item{x}{Nucleotide sequence data of virus isolates in fasta format.}

\item{serotyp}{Serotype information of the virus isolates whose sequence information was supplied in fasta format.}

\item{ind}{vector of two-length whose elements reprsent proportion of data used for model training and testing.}

\item{m}{Integral number (>0) indicating the number of runs performed on test data to evaluate the trained model.}

\item{k}{Integral number (>0) indicating the size of mers to generate features.}

\item{n}{number of features to be considered in model building.}
}
\value{
Returns the prediction model along with model fitting measures including mean and standard error in accuracy and prediction error over the runs on test data.
}
\description{
Returns the trained model along with results on test data.
}
\author{
Samarendra Das
}
