% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMDVSerPred.R
\name{serotypPredTest}
\alias{serotypPredTest}
\title{Testing of trained model.}
\usage{
serotypPredTest(x, serotyp, QuerySeq, serotypTest)
}
\arguments{
\item{x}{Nucleotide sequence data of virus isolates in fasta format for model training.}

\item{serotyp}{vector (size equal to number of virus isolates in x) containing serotype information of virus in x.}

\item{QuerySeq}{Sequnce data of of virus isolates in fasta format used for testing the trained model.}

\item{serotypTest}{vector (size equal to QuerySeq length) containing the serotype information of the isolates of QuerySeq.}
}
\value{
List containing predicted serotypes of isolates, Confusion matrix for trained model testing, accuracy of the trained model on test data.
\itemize{
  \item1 PredictedSerotypes conatins the serotype of isolates in TestSeq data using the trained model.
  \item2 ConfusionMatrix confusion matrix constructed by predicted serotypes vs. actual serotypes of the isolates in TestSea data.
  \item3 Accuracy is the trained model accuracy measure computed from the test
 }
}
\description{
Validates the trained model on independent test data and return performance metrics.
}
\author{
Samarendra Das
}
